change rate to big decimal and format rate

+++ app/src/main/java/com/chicohan/currencyexchange/data/model/DetailFragmentModel.kt
@@ -2,11 +2,12 @@ package com.chicohan.currencyexchange.data.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class DetailFragmentModel(
     val currentCurrency: String,
     val currentCurrencyAmount: Double,
     val calculatedCurrency: String,
-    val calculatedCurrencyAmount: Double
+    val calculatedCurrencyAmount: BigDecimal
 ): Parcelable
